backend/pyproject.toml
[tool.black]
line-length = 120
target_version = ['py310']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | versions
  )/
)
'''


[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["bekzod-fayzikuloff <fayzikuloff.b@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.4"
djangorestframework = "^3.13.1"
Markdown = "^3.3.6"
django-filter = "^21.1"
python-dotenv = "^0.20.0"
drf-yasg = {extras = ["validation"], version = "^1.20.0"}
psycopg2-binary = "^2.9.3"
django-cors-headers = "^3.11.0"
celery = "^5.2.6"
redis = "^4.3.1"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
isort = "^5.10.1"
pylint = "^2.13.7"
pre-commit = "^2.18.1"
flake8 = "^4.0.1"
Faker = "^13.11.0"
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
